# Pack store lru on offsets