--- a/bench/irmin-pack/tree.ml
+++ b/bench/irmin-pack/tree.ml
@@ -36,12 +36,21 @@ type config = {
   keep_stat_trace : bool;
   no_summary : bool;
   empty_blobs : bool;
+  gc_every : int;
+  gc_distance_in_the_past : int;
+  gc_wait_after : int;
 }
 
 module type Store = sig
   type store_config = config
+  type key
 
-  include Irmin.Generic_key.KV with type Schema.Contents.t = bytes
+  include
+    Irmin.Generic_key.KV
+      with type Schema.Contents.t = bytes
+       and type commit_key = key
+       and type node_key = key
+       and type contents_key = key
 
   type on_commit := int -> Hash.t -> unit Lwt.t
   type on_end := unit -> unit Lwt.t
@@ -157,6 +166,9 @@ module Bench_suite (Store : Store) = struct
         keep_stat_trace = config.keep_stat_trace;
         empty_blobs = config.empty_blobs;
         return_type = Summary;
+        gc_every = config.gc_every;
+        gc_distance_in_the_past = config.gc_distance_in_the_past;
+        gc_wait_after = config.gc_wait_after;
       }
     in
     if config.no_summary then
@@ -187,6 +199,8 @@ struct
   end
 
   include Store
+
+  type key = commit_key
 
   let indexing_strategy = Irmin_pack.Indexing_strategy.minimal
 
@@ -330,7 +344,7 @@ let get_suite suite_filter =
 let main () ncommits number_of_commits_to_replay suite_filter inode_config
     store_type freeze_commit path_conversion depth width nchain_trees
     nlarge_trees replay_trace_path artefacts_path keep_store keep_stat_trace
-    no_summary empty_blobs =
+    no_summary empty_blobs gc_every gc_distance_in_the_past gc_wait_after =
   let default = match suite_filter with `Quick -> 10000 | _ -> 13315 in
   let number_of_commits_to_replay =
     Option.value ~default number_of_commits_to_replay
@@ -354,6 +368,9 @@ let main () ncommits number_of_commits_to_replay suite_filter inode_config
       keep_stat_trace;
       no_summary;
       empty_blobs;
+      gc_every;
+      gc_distance_in_the_past;
+      gc_wait_after;
     }
   in
   Printexc.record_backtrace true;
@@ -513,6 +530,27 @@ let artefacts_path =
 let setup_log =
   Term.(const setup_log $ Fmt_cli.style_renderer () $ Logs_cli.level ())
 
+let gc_every =
+  let doc = Arg.info ~doc:"Distance between calls to GC" [ "gc_every" ] in
+  Arg.(value @@ opt int 1000 doc)
+
+let gc_distance_in_the_past =
+  let doc =
+    Arg.info ~doc:"Distance between the GC commit and the latest commit"
+      [ "gc-distance-in-the-past" ]
+  in
+  Arg.(value @@ opt int 5000 doc)
+
+let gc_wait_after =
+  let doc =
+    Arg.info
+      ~doc:
+        "How many commits separate the start of a GC and the moment we wait \
+         for the end of it"
+      [ "gc-wait-after" ]
+  in
+  Arg.(value @@ opt int 0 doc)
+
 let main_term =
   Term.(
     const main
@@ -533,7 +571,10 @@ let main_term =
     $ keep_store
     $ keep_stat_trace
     $ no_summary
-    $ empty_blobs)
+    $ empty_blobs
+    $ gc_every
+    $ gc_distance_in_the_past
+    $ gc_wait_after)
 
 let deprecated_info = (Term.info [@alert "-deprecated"])
 let deprecated_exit = (Term.exit [@alert "-deprecated"])