On top of #1922

Fixes #1844

Fixes #1920

This will have to be released as a 3.3.1 because the former code could create patterns in the store that will cause crash after GC.

@icristescu Is working on a regression test. 

A replay of 10k commits from genesis shows that the number of cache_miss stays roughly equivalent.

```
 |                          |  before  |     after
 | --                       | --       | --
 | -- main metrics --       |          | 
 | CPU time elapsed         |    1m24s |    1m24s 100%
 | Wall time elapsed        |    1m34s |    1m27s  92%
 | TZ-transactions per sec  |  106.083 |  106.486 100%
 | TZ-operations per sec    | 1196.450 | 1200.996 100%
 | Context.set per sec      | 8548.792 | 8581.276 100%
 | tail latency (1)         |  3.216 s |  2.181 s  68%
 |                          |          | 
 | -- resource usage --     |          | 
 | disk IO (total)          |          | 
 |   IOPS (op/sec)          |   27_936 |   28_554 102%
 |   throughput (bytes/sec) |  2.435 M |  2.469 M 101%
 |   total (bytes)          |  0.205 G |  0.207 G 101%
 | disk IO (read)           |          | 
 |   IOPS (op/sec)          |   27_327 |   27_943 102%
 |   throughput (bytes/sec) |  1.250 M |  1.279 M 102%
 |   total (bytes)          |  0.105 G |  0.107 G 102%
 | disk IO (write)          |          | 
 |   IOPS (op/sec)          |      609 |      611 100%
 |   throughput (bytes/sec) |  1.185 M |  1.189 M 100%
 |   total (bytes)          |  0.100 G |  0.100 G 100%
 |                          |          | 
 | max memory usage         |  1.001 G |  1.001 G 100%
 | mean CPU usage           |      89% |      97%     

 - (1) Longest Context.commit.
 - The "per sec" stats are calculated over CPU time.
 - "TZ-transactions" and "TZ-operations" are approximations.
 - "max memory usage" is the max size of OCaml's major heap.
 - "mean CPU usage" is inexact.

-- global --
 |                              |        | min per block | max per block  | avg per block |  avg per sec
 |                              |        |               |                |               | 
 | Add count                    | before |        3      |   123_869      |   71.987      |  7650.628     
 |                              | after  |        3 100% |   123_869 100% |   71.987 100% |  8293.185 108%
 | Remove count                 | before |        0      |    15_385      |    4.684      |   497.826     
 |                              | after  |        0      |    15_385 100% |    4.684 100% |   539.637 108%
 | Find count                   | before |        2      |    76_663      |  225.012      | 23913.687     
 |                              | after  |        2 100% |    76_663 100% |  225.012 100% | 25922.136 108%
 | Mem count                    | before |        0      |    93_342      |   53.579      |  5694.206     
 |                              | after  |        0      |    93_342 100% |   53.579 100% |  6172.448 108%
 | Mem_tree count               | before |        0      |        33      |   32.993      |  3506.460     
 |                              | after  |        0      |        33 100% |   32.993 100% |  3800.959 108%
 | Copy count                   | before |        0      |         7      |    0.005      |     0.489     
 |                              | after  |        0      |         7 100% |    0.005 100% |     0.530 108%
 | Commit count                 | before |        1      |         1      |    1.000      |   106.278     
 |                              | after  |        1 100% |         1 100% |    1.000 100% |   115.204 108%
 |                              |        |               |                |               | 
 | Disk bytes read              | before |        0      |   2.405 M      |  0.011 M      |   1.119 M     
 |                              | after  |        0      |   2.405 M 100% |  0.011 M 102% |   1.237 M 111%
 | Disk bytes written           | before |  0.000 M      |  13.997 M      |  0.010 M      |   1.060 M     
 |                              | after  |  0.000 M 100% |  13.997 M 100% |  0.010 M 100% |   1.149 M 108%
 | Disk bytes both              | before |  0.000 M      |  14.722 M      |  0.021 M      |   2.179 M     
 |                              | after  |  0.000 M 100% |  14.722 M 100% |  0.021 M 101% |   2.386 M 109%
 |                              |        |               |                |               | 
 | Disk reads                   | before |        0      |    53_452      |  230.118      | 24456.340     
 |                              | after  |        0      |    53_452 100% |  234.410 102% | 27004.923 110%
 | Disk writes                  | before |        4      |        54      |    5.126      |   544.779     
 |                              | after  |        4 100% |        54 100% |    5.126 100% |   590.534 108%
 | Disk both                    | before |        4      |    53_460      |  235.244      | 25001.119     
 |                              | after  |        4 100% |    53_460 100% |  239.536 102% | 27595.457 110%
 |                              |        |               |                |               | 
 | pack.finds                   | before |        0      |    16_619      |  518.010      | 55052.880     
 |                              | after  |        0      |    16_619 100% |  518.010 100% | 59676.631 108%
 | pack.finds.from_staging      | before |        0      |         0      |        0      |         0     
 |                              | after  |        0      |         0      |        0      |         0     
 | pack.finds.from_lru          | before |        0      |    15_901      |  499.783      | 53115.790     
 |                              | after  |        0      |    15_901 100% |  498.156 100% | 57389.321 108%
 | pack.finds.from_pack_direct  | before |        0      |     7_025      |   18.227      |  1937.090     
 |                              | after  |        0      |     7_308 104% |   19.855 109% |  2287.310 118%
 | pack.finds.from_pack_indexed | before |        0      |         0      |        0      |         0     
 |                              | after  |        0      |         0      |        0      |         0     
 | pack.finds.missing           | before |        0      |         0      |        0      |         0     
 |                              | after  |        0      |         0      |        0      |         0     
 | pack.finds.cache_miss        | before |        0      |     7_025      |   18.227      |  1937.090     
 |                              | after  |        0      |     7_308 104% |   19.855 109% |  2287.310 118%
 | pack.appended_hashes         | before |        0      |         0      |        0      |         0     
 |                              | after  |        0      |         0      |        0      |         0     
 | pack.appended_offsets        | before |        2      |   251_296      |  424.821      | 45148.922     
 |                              | after  |        2 100% |   251_296 100% |  424.821 100% | 48940.865 108%
 | pack.inode_add               | before |        0      |     7_029      |   81.662      |  8678.822     
 |                              | after  |        0      |     7_029 100% |   81.662 100% |  9407.734 108%
 | pack.inode_remove            | before |        0      |       208      |    0.968      |   102.824     
 |                              | after  |        0      |       208 100% |    0.968 100% |   111.459 108%
 | pack.inode_of_seq            | before |        0      |   166_816      |   26.099      |  2773.697     
 |                              | after  |        0      |   166_816 100% |   26.099 100% |  3006.653 108%
 | pack.inode_of_raw            | before |        0      |     1_760      |  238.518      | 25349.072     
 |                              | after  |        0      |     1_760 100% |  238.518 100% | 27478.076 108%
 | pack.inode_rec_add           | before |        0      |    91_732      |   97.034      | 10312.596     
 |                              | after  |        0      |    91_732 100% |   97.034 100% | 11178.724 108%
 | pack.inode_rec_remove        | before |        0      |       242      |    1.206      |   128.150     
 |                              | after  |        0      |       242 100% |    1.206 100% |   138.913 108%
 | pack.inode_to_binv           | before |        0      |   210_370      |  109.394      | 11626.113     
 |                              | after  |        0      |   210_370 100% |  109.394 100% | 12602.560 108%
 | pack.inode_decode_bin        | before |        0      |     7_025      |   18.197      |  1933.944     
 |                              | after  |        0      |     7_308 104% |   19.299 106% |  2223.291 115%
 | pack.inode_encode_bin        | before |        0      |   188_594      |   99.781      | 10604.477     
 |                              | after  |        0      |   188_594 100% |   99.781 100% | 11495.119 108%
 |                              |        |               |                |               | 
 | tree.contents_hash           | before |        0      |    62_693      |   28.781      |  3058.744     
 |                              | after  |        0      |    62_693 100% |   28.781 100% |  3315.640 108%
 | tree.contents_find           | before |        0      |       749      |   93.497      |  9936.617     
 |                              | after  |        0      |       749 100% |   93.497 100% | 10771.168 108%
 | tree.contents_add            | before |        0      |    62_693      |   28.764      |  3057.012     
 |                              | after  |        0      |    62_693 100% |   28.764 100% |  3313.763 108%
 | tree.node_hash               | before |        0      |         0      |        0      |         0     
 |                              | after  |        0      |         0      |        0      |         0     
 | tree.node_mem                | before |        0      |         0      |        0      |         0     
 |                              | after  |        0      |         0      |        0      |         0     
 | tree.node_add                | before |        0      |   166_818      |   89.480      |  9509.764     
 |                              | after  |        0      |   166_818 100% |   89.480 100% | 10308.465 108%
 | tree.node_find               | before |        0      |     1_760      |  238.518      | 25349.072     
 |                              | after  |        0      |     1_760 100% |  238.518 100% | 27478.076 108%
 | tree.node_val_v              | before |        0      |   166_816      |   26.099      |  2773.718     
 |                              | after  |        0      |   166_816 100% |   26.099 100% |  3006.676 108%
 | tree.node_val_find           | before |        0      |         0      |        0      |         0     
 |                              | after  |        0      |         0      |        0      |         0     
 | tree.node_val_list           | before |        0      |       326      |    4.428      |   470.587     
 |                              | after  |        0      |       326 100% |    4.428 100% |   510.110 108%
 |                              |        |               |                |               | 
 | index.cumu_data_bytes        | before |        0      |         0      |        0      |         0     
 |                              | after  |        0      |         0      |        0      |         0     
 |                              |        |               |                |               | 
 | gc.minor_words allocated     | before |  0.018 M      | 630.138 M      |  0.882 M      |  93.713 M     
 |                              | after  |  0.018 M 100% | 626.431 M  99% |  0.874 M  99% | 100.641 M 107%
 | gc.major_words allocated     | before |        0      |  45.608 M      |  0.063 M      |   6.695 M     
 |                              | after  |        0      |  45.819 M 100% |  0.063 M 100% |   7.229 M 108%
 | gc.minor_collections         | before |        0      |     2_411      |    3.367      |   357.794     
 |                              | after  |        0      |     2_397  99% |    3.335  99% |   384.239 107%
 | gc.major_collections         | before |        0      |        13      |    0.006      |     0.638     
 |                              | after  |        0      |        13 100% |    0.006  95% |     0.657 103%

 |                               |        |      min      |       max        |      avg
 |                               |        |               |                  | 
 | Block duration (s)            | before | 0.813 ms      |     5.338 s      |  9.409 ms     
 |                               | after  | 0.731 ms  90% |     3.904 s  73% |  8.680 ms  92%
 | Buildup duration (s)          | before | 0.389 ms      |     2.122 s      |  5.133 ms     
 |                               | after  | 0.382 ms  98% |     1.723 s  81% |  4.405 ms  86%
 | Commit duration (s)           | before | 0.117 ms      |     3.216 s      |  4.277 ms     
 |                               | after  | 0.104 ms  89% |     2.181 s  68% |  4.275 ms 100%
 |                               |        |               |                  | 
 | Add duration (s)              | before | 0.478 µs      |  322.171 ms      |  9.508 µs     
 |                               | after  | 0.472 µs  99% |  401.193 ms 125% |  9.656 µs 102%
 | Remove duration (s)           | before | 0.986 µs      |   39.159 ms      | 12.718 µs     
 |                               | after  | 1.002 µs 102% |    9.407 ms  24% |  8.757 µs  69%
 | Find duration (s)             | before | 0.223 µs      |   14.324 ms      | 11.127 µs     
 |                               | after  | 0.222 µs 100% | 1430.964 ms 100x |  9.369 µs  84%
 | Mem duration (s)              | before | 0.409 µs      | 1419.649 ms      | 13.400 µs     
 |                               | after  | 0.411 µs 100% |    6.984 ms 0.5% |  6.989 µs  52%
 | Mem_tree duration (s)         | before | 0.505 µs      |    5.436 ms      |  2.912 µs     
 |                               | after  | 0.511 µs 101% |    4.947 ms  91% |  3.010 µs 103%
 | Copy duration (s)             | before | 2.445 µs      |    1.354 ms      | 41.869 µs     
 |                               | after  | 2.291 µs  94% |    0.184 ms  14% | 15.749 µs  38%
 | Unseen duration (s)           | before | 0.363 ms      |     1.164 s      |  1.064 ms     
 |                               | after  | 0.340 ms  94% |     0.960 s  82% |  1.080 ms 102%
 |                               |        |               |                  | 
 | Major heap bytes after commit | before |  9.138 M      |  1000.641 M      | 523.847 M     
 |                               | after  |  9.138 M 100% |  1000.641 M 100% | 552.942 M 106%
 |                               |        |               |                  | 
 | CPU Usage                     | before |      12%      |        103%      |       94%     
 |                               | after  |      10%      |        100%      |       98%     

-- evolution through blocks --
 | Block played count *C                      |        | 0 (before) |      3350       |      6700       |   10000 (end)
 | Blocks progress *C                         |        |     0%     |       34%       |       67%       |      100%
 |                                            |        |            |                 |                 | 
 | Wall time elapsed *C                       | before |        0us |    32.253s      |     1min3s      |    1min34s     
 |                                            | after  |        0us |    23.586s  73% |    53.419s  84% |    1min26s  92%
 | CPU time elapsed *C                        | before |        0us |    29.394s      |    56.693s      |    1min24s     
 |                                            | after  |        0us |    23.544s  80% |    52.347s  92% |    1min23s 100%
 | CPU Usage *LA                              | before |        n/a |        92%      |        90%      |        98%     
 |                                            | after  |        n/a |       100%      |        98%      |        96%     
 |                                            |        |            |                 |                 | 
 | Approx. TZ-transaction count *C            | before |          0 |      2_080      |      5_532      |      8_933     
 |                                            | after  |          0 |      2_080 100% |      5_532 100% |      8_933 100%
 | Approx. TZ-operations count *C             | before |          0 |     32_224      |     66_745      |    100_750     
 |                                            | after  |          0 |     32_224 100% |     66_745 100% |    100_750 100%
 | Op count *C                                | before |          0 |  1_518_984      |  2_752_540      |  3_902_594     
 |                                            | after  |          0 |  1_518_984 100% |  2_752_540 100% |  3_902_594 100%
 |                                            |        |            |                 |                 | 
 | Commit per sec *LA *N                      | before |        n/a |    113.866      |     91.756      |    110.661     
 |                                            | after  |        n/a |    131.845 116% |    101.028 110% |    100.642  91%
 | Add per sec *LA *N                         | before |        n/a |   7724.495      |   5182.417      |   5639.293     
 |                                            | after  |        n/a |   8944.142 116% |   5706.138 110% |   5128.724  91%
 | Remove per sec *LA *N                      | before |        n/a |    467.165      |    273.327      |    182.001     
 |                                            | after  |        n/a |    540.927 116% |    300.948 110% |    165.523  91%
 | Find per sec *LA *N                        | before |        n/a |  26338.254      |  19585.191      |  22949.051     
 |                                            | after  |        n/a |  30496.892 116% |  21564.414 110% |  20871.295  91%
 | Mem per sec *LA *N                         | before |        n/a |   6333.563      |   3859.956      |   4060.901     
 |                                            | after  |        n/a |   7333.592 116% |   4250.032 110% |   3693.236  91%
 | Mem_tree per sec *LA *N                    | before |        n/a |   3757.594      |   3027.942      |   3651.822     
 |                                            | after  |        n/a |   4350.893 116% |   3333.937 110% |   3321.194  91%
 | Copy per sec *LA *N                        | before |        n/a |      0.739      |      0.527      |      0.743     
 |                                            | after  |        n/a |      0.856 116% |      0.581 110% |      0.675  91%
 |                                            |        |            |                 |                 | 
 | Add count per block *LA                    | before |        n/a |     67.838      |     56.481      |     50.960     
 |                                            | after  |        n/a |     67.838 100% |     56.481 100% |     50.960 100%
 | Remove count per block *LA                 | before |        n/a |      4.103      |      2.979      |      1.645     
 |                                            | after  |        n/a |      4.103 100% |      2.979 100% |      1.645 100%
 | Find count per block *LA                   | before |        n/a |    231.308      |    213.449      |    207.381     
 |                                            | after  |        n/a |    231.308 100% |    213.449 100% |    207.381 100%
…