@Ngoguey42 2022-06-22
```
 |                              |        | min per block | max per block  | avg per block |  avg per sec
 |                              |        |               |                |               | 
 | pack.finds.from_lru          | before |        0      |    15_901      |  499.783      | 53115.790     
 |                              | after  |        0      |    15_901 100% |  498.156 100% | 57389.321 108%
```

It decreases from 499.783 per block to 498.156 per block